Fixing native driver build in Windows and bugfixes
- added prebuilt dependencies to github thrift SnappyData fork and mechanism
  in build to download and extract them; use these instead of hardcoded /export path
- removed unused CallStack (will move to CppCMS backtrace that works on all platforms
    later: https://sourceforge.net/p/cppcms/code/HEAD/tree/framework/trunk/booster/lib/backtrace)
- added Visual Studio solution generation in build.gradle and prebuilt solution for convenience
- many fixes for windows build

+++ native/src/snappyclient/cpp/UpdatableRow.cpp
@@ -68,7 +68,7 @@ std::vector<int32_t> UpdatableRow::getChangedColumnsAsVector() {
   if (bitSet != NULL) {
     for (size_t pos = bitSet->find_first(); pos != DynamicBitSet::npos;
         pos = bitSet->find_next(pos)) {
-      changedColumns.push_back(pos);
+      changedColumns.push_back(static_cast<int32_t>(pos));
     }
   }
   return std::move(changedColumns);
